Adding last bit of unit tests, cant seem to get the drvier working
