+++ Website/Main/Views/Home/Index.cshtml
@@ -82,7 +82,7 @@
           <div class="item">
             <img src="assets/images/course-01.jpg" alt="Course One" width="400" height="200" >
             <div class="down-content">
-                    <input class = "IvansText" type="submit" value="Lake Oswego, OR">
+                    <input id = "Lake Oswego, OR" class = "IvansText" type="submit" value="Lake Oswego, OR">
             </div>
           </div>
 </form>
@@ -95,7 +95,7 @@
           <div class="item">
             <img src="assets/images/course-02.jpg" alt="Course Two" width="400" height="200">
             <div class="down-content">
-                    <input class = "IvansText" type="submit" value="Sedona, AZ">
+                    <input id = "Sedona, AZ" class = "IvansText" type="submit" value="Sedona, AZ">
             </div>
           </div>
 </form>
@@ -108,7 +108,7 @@
           <div class="item">
             <img src="assets/images/course-03.jpg" alt="" width="400" height="200">
             <div class="down-content">
-                    <input class = "IvansText" type="submit" value="Shaker Heights, OH">
+                    <input id = "Shaker Heights, OH" class = "IvansText" type="submit" value="Shaker Heights, OH">
             </div>
           </div>
 </form>
@@ -122,7 +122,7 @@
           <div class="item">
             <img src="assets/images/GreenWood.jpg" alt="" width="400" height="200">
             <div class="down-content">
-                    <input class = "IvansText" type="submit" value="Greenwood Village, CO">
+                    <input id = "Greenwood Village, CO" class = "IvansText" type="submit" value="Greenwood Village, CO">
             </div>
           </div>
 </form>
@@ -136,7 +136,7 @@
           <div class="item">
             <img src="assets/images/Evanston.jpg" alt="" width="400" height="200">
             <div class="down-content">
-                    <input class = "IvansText" type="submit" value="Evanston, IL">
+                    <input id = "tests" class = "IvansText" type="submit" value="Evanston, IL">
             </div>
             </div>
 </form>
@@ -150,7 +150,7 @@
           <div class="item">
             <img src="assets/images/Amherst.jpg" alt="" width="400" height="200">
             <div class="down-content">
-                    <input class = "IvansText" type="submit" value="Amherst, NH">
+                    <input id = "tests" class = "IvansText" type="submit" value="Amherst, NH">
             </div>
           </div>
 </form>
